refactor: use a util function to get the transfer AssetInfo
