+++ src/routes/safe/components/Transactions/TxList/TxHistoryCollapsed.tsx
@@ -1,8 +1,8 @@
 import { MultisigExecutionInfo } from '@gnosis.pm/safe-react-gateway-sdk'
-import { ReactElement } from 'react'
+import { ReactElement, useMemo } from 'react'
 
 import { Transaction } from 'src/logic/safe/store/models/types/gateway.d'
-import { useAssetInfo } from './hooks/useAssetInfo'
+import { getAssetInfo } from 'src/routes/safe/components/Transactions/TxList/utils'
 import { useTransactionStatus } from './hooks/useTransactionStatus'
 import { useTransactionType } from './hooks/useTransactionType'
 import { TxCollapsed } from './TxCollapsed'
@@ -10,7 +10,7 @@ import { TxCollapsed } from './TxCollapsed'
 export const TxHistoryCollapsed = ({ transaction }: { transaction: Transaction }): ReactElement => {
   const nonce = (transaction.executionInfo as MultisigExecutionInfo)?.nonce
   const type = useTransactionType(transaction)
-  const info = useAssetInfo(transaction.txInfo)
+  const info = useMemo(() => getAssetInfo(transaction.txInfo), [transaction.txInfo])
   const status = useTransactionStatus(transaction)
 
   return (
