+++ src/routes/safe/components/Transactions/TxList/TxQueueCollapsed.tsx
@@ -1,9 +1,9 @@
 import { MultisigExecutionInfo } from '@gnosis.pm/safe-react-gateway-sdk'
-import { ReactElement } from 'react'
+import { ReactElement, useMemo } from 'react'
 
 import useTxStatus from 'src/logic/hooks/useTxStatus'
 import { LocalTransactionStatus, Transaction } from 'src/logic/safe/store/models/types/gateway.d'
-import { useAssetInfo } from './hooks/useAssetInfo'
+import { getAssetInfo } from 'src/routes/safe/components/Transactions/TxList/utils'
 import { useTransactionStatus } from './hooks/useTransactionStatus'
 import { useTransactionType } from './hooks/useTransactionType'
 import { TxCollapsed } from './TxCollapsed'
@@ -34,7 +34,7 @@ export const TxQueueCollapsed = ({ isGrouped = false, transaction }: TxQueuedCol
   const executionInfo = transaction.executionInfo as MultisigExecutionInfo
   const nonce = executionInfo?.nonce
   const type = useTransactionType(transaction)
-  const info = useAssetInfo(transaction.txInfo)
+  const info = useMemo(() => getAssetInfo(transaction.txInfo), [transaction.txInfo])
   const status = useTransactionStatus(transaction)
   const txStatus = useTxStatus(transaction)
   const isPending = txStatus === LocalTransactionStatus.PENDING
